fixed bug with navigation on elements with a dot in their name
The bug occured when a navigation step with an element name with a dot
in it (eg elt.foo) was called before an element by that name had been
seen in the XML. At this point the code tried to treat the expression
as an element with a class attribute that contained "foo", like in CSS
selectors, and thus did not find the proper element.

The fix is for XML::Twig to look for either a litteral element by that
name _or_ an element with the appropriate class.

Additionally, using css_sel => 0 forces _not_ treating the dot as the
class separator.